Add score count DOM updates and logic for final winner to playRound func
The playRound function looks at the scores and if any of them are > 4 the game stops and the final winner is shown. Scores are updated based on the outcome of the determineWinner function that returns an array (the array's values depend on who won the round array[0] contains the human score and array[1] containsthe computer score; draw returns [0,0]).